+++ packages/bezier-react/src/components/AlphaStack/index.ts
@@ -1,0 +1,2 @@
+export { AlphaStack } from './AlphaStack'
+export type { AlphaStackProps } from './AlphaStack.types'