+++ packages/bezier-react/src/components/Avatars/Avatar/Avatar.tsx
@@ -12,7 +12,9 @@ import {
   Status,
   StatusSize,
 } from '~/src/components/Status'
-import defaultAvatarUrl from '~/src/components/Avatars/assets/defaultAvatar.svg'
+// NOTE: Don't fix it. When using absolute paths, a module resolve error occurs at the runtime of the library consumer.
+// eslint-disable-next-line no-restricted-imports
+import defaultAvatarUrl from '../assets/defaultAvatar.svg'
 import useProgressiveImage from './useProgressiveImage'
 import type AvatarProps from './Avatar.types'
 import { AvatarSize } from './Avatar.types'