+++ packages/bezier-react/src/components/Forms/Checkbox/Checkbox.stories.tsx
@@ -8,7 +8,10 @@ import {
 
 /* Internal dependencies */
 import { getTitle } from '~/src/utils/storyUtils'
-import { type CheckboxProps } from './Checkbox.types'
+import {
+  type CheckboxProps,
+  type CheckedState,
+} from './Checkbox.types'
 import { Checkbox } from './Checkbox'
 
 export default {
@@ -28,7 +31,7 @@ export default {
   },
 } as Meta
 
-const Template: Story<CheckboxProps> = ({ children, ...otherCheckboxProps }) => (
+const Template: Story<CheckboxProps<CheckedState>> = ({ children, ...otherCheckboxProps }) => (
   <Checkbox {...otherCheckboxProps}>
     { children }
   </Checkbox>