+++ packages/bezier-react/src/components/Forms/Checkbox/Checkbox.types.ts
@@ -7,17 +7,17 @@ import { type FormComponentProps } from '~/src/components/Forms'
 
 export type CheckedState = boolean | 'indeterminate'
 
-interface CheckboxOptions {
+interface CheckboxOptions<Checked extends CheckedState> {
   /**
    * The controlled checked state of the checkbox.
    * Must be used in conjunction with `onCheckedChange`.
    */
-  checked?: CheckedState
+  checked?: Checked
   /**
    * The checked state of the checkbox when it is initially rendered.
    * Use when you do not need to control its checked state.
    */
-  defaultChecked?: CheckedState
+  defaultChecked?: Checked
   /**
    * The unique id of the checkbox. It is created automatically by default.
    * It used by the label element in the checkbox.
@@ -36,12 +36,12 @@ interface CheckboxOptions {
   /**
    * Event handler called when the checked state of the checkbox changes.
    */
-  onCheckedChange?: (checked: CheckedState) => void
+  onCheckedChange?: (checked: Checked) => void
 }
 
-export interface CheckboxProps extends
+export interface CheckboxProps<Checked extends CheckedState> extends
   Omit<BezierComponentProps, 'as'>,
   ChildrenProps,
   FormComponentProps,
-  Omit<React.HTMLAttributes<HTMLButtonElement>, keyof CheckboxOptions>,
-  CheckboxOptions {}
+  Omit<React.HTMLAttributes<HTMLButtonElement>, keyof CheckboxOptions<Checked>>,
+  CheckboxOptions<Checked> {}