+++ packages/bezier-react/src/foundation/Colors/Palette/index.ts
@@ -119,11 +119,13 @@ type BaseColorfulPaletteKey300_20 = `${BaseColorfulPaletteKey300}_20`
 
 type AlphaGreyKey =
   | `${BasePaletteKey.Grey}900_90`
+  | `${BasePaletteKey.Grey}850_80`
   | `${BasePaletteKey.Grey}800_80`
   | `${BasePaletteKey.Grey}700_80`
   | `${BasePaletteKey.Grey}200_80`
   | `${BasePaletteKey.Grey}100_90`
   | `${BasePaletteKey.Grey}100_80`
+  | `${BasePaletteKey.Grey}50_80`
 
 type AlphaWhiteKey =
   | `${BasePaletteKey.White}_90`
@@ -254,11 +256,13 @@ export const Palette: PaletteType = {
 
   // Alpha Grey
   grey900_90: getAlphaHex(PaletteWithoutAlpha.grey900, 90),
+  grey850_80: getAlphaHex(PaletteWithoutAlpha.grey850, 80),
   grey800_80: getAlphaHex(PaletteWithoutAlpha.grey800, 80),
   grey700_80: getAlphaHex(PaletteWithoutAlpha.grey700, 80),
   grey200_80: getAlphaHex(PaletteWithoutAlpha.grey200, 80),
   grey100_90: getAlphaHex(PaletteWithoutAlpha.grey100, 90),
   grey100_80: getAlphaHex(PaletteWithoutAlpha.grey100, 80),
+  grey50_80: getAlphaHex(PaletteWithoutAlpha.grey50, 80),
 
   // Alpha White
   white_90: getAlphaHex(PaletteWithoutAlpha.white, 90),